use platform.chrome.allservices.redesign flag
