+++ src/layouts/AllServices.tsx
@@ -20,6 +20,8 @@ import useAllServices from '../hooks/useAllServices';
 import Messages from '../locales/Messages';
 import { updateDocumentTitle } from '../utils/common';
 import fetchNavigationFiles from '../utils/fetchNavigationFiles';
+import { useFlag } from '@unleash/proxy-client-react';
+import AllServicesBundle from '../components/AllServices/AllServicesBundle';
 
 const availableBundles = ['openshift', 'insights', 'ansible', 'settings', 'iam', 'subscriptions'];
 
@@ -31,6 +33,8 @@ export type AllServicesProps = {
 };
 
 const AllServices = ({ Footer }: AllServicesProps) => {
+  const enableAllServicesRedesign = useFlag('platform.chrome.allservices.redesign');
+
   updateDocumentTitle('All Services', true);
   const { linkSections, error, ready, filterValue, setFilterValue } = useAllServices();
   const intl = useIntl();
@@ -92,9 +96,9 @@ const AllServices = ({ Footer }: AllServicesProps) => {
             </PageGroup>
             <PageSection hasBodyWrapper={false} padding={{ default: 'noPadding', md: 'padding', lg: 'padding' }} className="pf-v6-u-pt-lg">
               <Gallery className="pf-v6-u-display-block" hasGutter>
-                {filteredBundles.map((bundle, index) => (
-                  <AllServicesSection key={index} {...bundle} />
-                ))}
+                {enableAllServicesRedesign
+                  ? filteredBundles.map((bundle, index) => <AllServicesBundle key={index} {...bundle} />)
+                  : sections.map((section, index) => <AllServicesSection key={index} {...section} />)}
                 {/* TODO: Add empty state */}
                 {sections.length === 0 && filterValue.length !== 0 && <div>Nothing found</div>}
               </Gallery>
